Ensure that the base is a 64-bit unsigned integer

+++ include/inform/time_series.h
@@ -22,7 +22,7 @@
  *
  * @see inform_active_info_ensemble
  */
-entropy inform_active_info(uint64_t const *series, size_t n, int base, uint64_t k);
+entropy inform_active_info(uint64_t const *series, size_t n, uint64_t base, uint64_t k);
 
 /**
  * Compute the active information of an ensemble of time series.
@@ -43,4 +43,4 @@ entropy inform_active_info(uint64_t const *series, size_t n, int base, uint64_t 
  *
  * @see inform_active_info
  */
-entropy inform_active_info_ensemble(uint64_t const *series, size_t n, size_t m, int base, uint64_t k);
+entropy inform_active_info_ensemble(uint64_t const *series, size_t n, size_t m, uint64_t base, uint64_t k);
